// apps/admin-web/src/.generated/api/model/adminServiceTermGroupItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

export interface AdminServiceTermGroupItemDto {
  id: string;
  code: string;
  title: string;
  isRequired: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}